move ValidationTexts
