+++ src/ReactiveUI.Validation/Formatters/Abstractions/IValidationTextFormatter.cs
@@ -3,7 +3,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
-using ReactiveUI.Validation.Collections;
+using ReactiveUI.Validation.ValidationTexts.Abstractions;
 
 namespace ReactiveUI.Validation.Formatters.Abstractions;
