+++ src/ReactiveUI.Validation/Formatters/SingleLineFormatter.cs
@@ -3,8 +3,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
-using ReactiveUI.Validation.Collections;
 using ReactiveUI.Validation.Formatters.Abstractions;
+using ReactiveUI.Validation.ValidationTexts.Abstractions;
 
 namespace ReactiveUI.Validation.Formatters;
