+++ src/ReactiveUI.Validation/Helpers/ReactiveValidationObject.cs
@@ -13,11 +13,12 @@ using System.Reactive.Linq;
 using DynamicData;
 using DynamicData.Binding;
 using ReactiveUI.Validation.Abstractions;
-using ReactiveUI.Validation.Collections;
 using ReactiveUI.Validation.Components.Abstractions;
 using ReactiveUI.Validation.Contexts;
+using ReactiveUI.Validation.Contexts.Abstractions;
 using ReactiveUI.Validation.Formatters;
 using ReactiveUI.Validation.Formatters.Abstractions;
+using ReactiveUI.Validation.ValidationTexts;
 using Splat;
 
 namespace ReactiveUI.Validation.Helpers;
